chore: split tests rum enabled / disabled
